@@ -1,8 +1,16 @@
-cmake_minimum_required(VERSION 3.5)
-project(cstar)
+cmake_minimum_required(VERSION 3.28)
 
-# set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DMAC -std=c++11")
+include(FetchContent)
+set(FETCHCONTENT_UPDATES_DISCONNECTED TRUE)
+FetchContent_Declare(ftxui
+        GIT_REPOSITORY https://github.com/ArthurSonzogni/ftxui
+        GIT_TAG cdf28903a7781f97ba94d30b79c3a4b0c97ccce7
+)
+FetchContent_MakeAvailable(ftxui)
+
+project(cstar LANGUAGES CXX VERSION 3.0.0)
+
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DMAC -std=c++23 -stdlib=libc++")
 # set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -DDEBUG -g -ggdb -ggnu-pubnames -std=c++11")
 
 set(HEADERS cs_global.h cs_basic.h cs_block.h cs_compile.h cs_defines.h cs_errors.h
@@ -25,3 +33,9 @@ add_executable(cstar ${SOURCE_FILES} ${HEADERS})
 #add_executable(cstar main.cpp)
 #        commdelay.cpp
 #        interpret.cpp)
+
+target_link_libraries(cstar
+        PRIVATE ftxui::screen
+        PRIVATE ftxui::dom
+        PRIVATE ftxui::component
+)
