@@ -12,7 +12,7 @@
 #include "cs_defines.h"
 #include <vector>
 namespace Cstar {
-#define ArraySize(A) ((sizeof A) / (sizeof A[0]))
+#define ArraySize(A) ((sizeof(A)) / (sizeof(A)[0]))
 
 //    GLOBAL_CS_EXPORT std::istream *SRC;
 //    GLOBAL_CS_EXPORT std::ofstream *LIS;
@@ -37,8 +37,9 @@ struct ORDER {
   int F;
   int X;
   int Y;
-};
-enum OBJECTS {
+} __attribute__((aligned(16)));
+
+enum class OBJECTS : std::uint8_t {
   KONSTANT,
   VARIABLE,
   TYPE1,
@@ -53,12 +54,12 @@ enum OBJECTS {
 //        LOCKS, VOIDS
 //    };
 
-GLOBAL_CS_EXPORT int DISPLAY[LMAX + 1];
+GLOBAL_CS_EXPORT std::array<int, LMAX + 1> DISPLAY;
 
-typedef char ALFA[ALNG + 1];
-typedef char OLDALFA[OLDALNG];
-typedef char MESSAGEALFA[MESSAGELNG];
-typedef int INDEX;
+using ALFA = std::string;
+using OLDALFA = std::string;
+using MESSAGEALFA = std::string;
+using Index = int;
 
 #ifdef EXPORT_CS_GLOBAL
 GLOBAL_CS_EXPORT std::vector<int> LOCATION(LINELIMIT + 1);
@@ -83,12 +84,12 @@ GLOBAL_CS_EXPORT int BREAKPNT;
 #ifdef EXPORT_CS_GLOBAL
 GLOBAL_CS_EXPORT ORDER CODE[CMAX + 1];
 #else
-GLOBAL_CS_EXPORT ORDER CODE[CMAX + 1];
+GLOBAL_CS_EXPORT std::array<ORDER, CMAX + 1> code;
 #endif
 #ifdef EXPORT_CS_GLOBAL
 GLOBAL_CS_EXPORT std::vector<INDEX> WITHTAB(WMAX + 1);
 #else
-GLOBAL_CS_EXPORT std::vector<INDEX> WITHTAB;
+GLOBAL_CS_EXPORT std::vector<Index> WITHTAB;
 #endif
 GLOBAL_CS_EXPORT bool INCLUDEFLAG;
 GLOBAL_CS_EXPORT bool OKBREAK;
@@ -97,7 +98,7 @@ GLOBAL_CS_EXPORT int INUM;
 GLOBAL_CS_EXPORT double RNUM;
 GLOBAL_CS_EXPORT int SLENG;
 GLOBAL_CS_EXPORT char CH;
-GLOBAL_CS_EXPORT double CONTABLE[RCMAX + 1];
+GLOBAL_CS_EXPORT std::array<double, RCMAX + 1> CONTABLE;
 void EMIT(int FCT);
 void EMIT1(int FCT, int B);
 void EMIT2(int FCT, int A, int B);
