@@ -37,10 +37,10 @@ void INITCHAN(InterpLocal *il, int CHID) {
 int HOPS(int SOURCE, int DEST) {
   int T1, T2, T3, T4, DIST;
   DIST = 0;
-  switch (TOPOLOGY) {
-  case FULLCONNSY:
-  case CLUSTERSY:
-  case HYPERCUBESY:
+  switch (Topology) {
+  case Symbol::FULLCONNSY:
+  case Symbol::CLUSTERSY:
+  case Symbol::HYPERCUBESY:
     T1 = SOURCE;
     T2 = DEST;
     DIST = 0;
@@ -52,16 +52,16 @@ int HOPS(int SOURCE, int DEST) {
       T2 /= 2;
     }
     break;
-  case LINESY:
+  case Symbol::LINESY:
     DIST = abs(SOURCE - DEST);
     break;
-  case MESH2SY:
+  case Symbol::MESH2SY:
     T1 = TOPDIM;
     T2 = abs(SOURCE / T1 - DEST / T1);
     T3 = abs(SOURCE % T1 - DEST % T1);
     DIST = T2 + T3;
     break;
-  case MESH3SY:
+  case Symbol::MESH3SY:
     T2 = TOPDIM;
     T1 = TOPDIM * T2;
     T3 = SOURCE % T1;
@@ -69,7 +69,7 @@ int HOPS(int SOURCE, int DEST) {
     DIST = abs(T3 / T2 - T4 / T2) + abs(T3 % T2 - T4 % T2) +
            abs(SOURCE / T1 - DEST / T1);
     break;
-  case RINGSY:
+  case Symbol::RINGSY:
     T1 = abs(SOURCE - DEST);
     T2 = TOPDIM - T1;
     if (T1 < T2) {
@@ -78,7 +78,7 @@ int HOPS(int SOURCE, int DEST) {
       DIST = T2;
     }
     break;
-  case TORUSSY:
+  case Symbol::TORUSSY:
     T1 = TOPDIM;
     T3 = abs(SOURCE / T1 - DEST / T1);
     if (T3 > T1 / 2) {
@@ -101,28 +101,28 @@ int GRPDELAY(InterpLocal *il, int lSRC, int DEST, int LEN) {
   int AVEFAN = 0, MAXFAN = 0;
   NUMPACK = LEN / 3;
   NUMHOP = HOPS(lSRC, DEST);
-  switch (TOPOLOGY) {
-  case HYPERCUBESY:
+  switch (Topology) {
+  case Symbol::HYPERCUBESY:
     AVEFAN = TOPDIM;
     MAXFAN = TOPDIM;
     break;
-  case LINESY:
-  case RINGSY:
+  case Symbol::LINESY:
+  case Symbol::RINGSY:
     AVEFAN = 1;
     MAXFAN = 2;
     break;
-  case MESH2SY:
-  case TORUSSY:
+  case Symbol::MESH2SY:
+  case Symbol::TORUSSY:
     AVEFAN = 2;
     MAXFAN = 3;
     break;
-  case MESH3SY:
+  case Symbol::MESH3SY:
     AVEFAN = 3;
     MAXFAN = 4;
     break;
-  case FULLCONNSY:
-  case CLUSTERSY:
-    AVEFAN = (int)round(log(HIGHESTPROCESSOR + 1) / log(2));
+  case Symbol::FULLCONNSY:
+  case Symbol::CLUSTERSY:
+    AVEFAN = (int)round(log(HighestProcessor + 1) / log(2));
     MAXFAN = AVEFAN;
     break;
   default:
@@ -739,7 +739,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       break;
     }
     xl.H1 = il->S[CURPR->T];
-    il->S[xl.H1] = HIGHESTPROCESSOR + 1;
+    il->S[xl.H1] = HighestProcessor + 1;
     if (il->NUMTRACE > 0) {
       CHKVAR(il, xl.H1);
     }
@@ -752,7 +752,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       break;
     }
     if (CHKCOMM(&xl, il->S[CURPR->T]) || (il->S[CURPR->T - 1] < 0) ||
-        (il->S[CURPR->T - 2] < 0) || (il->S[CURPR->T - 2] > HIGHESTPROCESSOR) ||
+        (il->S[CURPR->T - 2] < 0) || (il->S[CURPR->T - 2] > HighestProcessor) ||
         (il->S[CURPR->T - 3] != MPIINT && il->S[CURPR->T - 3] != MPIFLOAT &&
          il->S[CURPR->T - 3] != MPICHAR)) {
       il->PS = InterpLocal::PS::MPIPARCHK;
@@ -852,7 +852,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->S[CURPR->T] <= 0 || CHKCOMM(&xl, il->S[CURPR->T - 1]) ||
         ((il->S[CURPR->T - 2] < 0) && (il->S[CURPR->T - 2] != MPIANYTAG)) ||
         ((il->S[CURPR->T - 3] < 0) && (il->S[CURPR->T - 3] != MPIANYSOURCE)) ||
-        (il->S[CURPR->T - 3] > HIGHESTPROCESSOR) ||
+        (il->S[CURPR->T - 3] > HighestProcessor) ||
         (il->S[CURPR->T - 4] != MPIINT && il->S[CURPR->T - 4] != MPIFLOAT &&
          il->S[CURPR->T - 4] != MPICHAR)) {
       il->PS = InterpLocal::PS::MPIPARCHK;
@@ -962,7 +962,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->NUMTRACE > 0) {
       CHKVAR(il, xl.H2);
     }
-    if (il->S[xl.H1] < 0 || il->S[xl.H1] > HIGHESTPROCESSOR ||
+    if (il->S[xl.H1] < 0 || il->S[xl.H1] > HighestProcessor ||
         il->S[xl.H1 + 1] < 0 || il->S[xl.H1 + 2] != 0) {
       il->PS = InterpLocal::PS::MPIPARCHK;
       break;
@@ -987,13 +987,13 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 57;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
     } else if (il->MPICODE != 57) {
       il->PS = InterpLocal::PS::MPIGRPCHK;
       break;
     }
     il->MPISEM = il->MPISEM - 1;
-    xl.H1 = HIGHESTPROCESSOR / 2;
+    xl.H1 = HighestProcessor / 2;
     if (CURPR->PID != xl.H1) {
       xl.R1 = CURPR->TIME + GRPDELAY(xl.il, CURPR->PROCESSOR, xl.H1, 1);
     } else {
@@ -1027,7 +1027,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       break;
     }
     if (CHKCOMM(&xl, il->S[CURPR->T]) || (il->S[CURPR->T - 1] < 0) ||
-        (il->S[CURPR->T - 1] > HIGHESTPROCESSOR) ||
+        (il->S[CURPR->T - 1] > HighestProcessor) ||
         (il->S[CURPR->T - 2] != MPIINT && il->S[CURPR->T - 2] != MPIFLOAT &&
          il->S[CURPR->T - 2] != MPICHAR)) {
       il->PS = InterpLocal::PS::MPIPARCHK;
@@ -1039,7 +1039,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 58;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
       il->MPIROOT = il->S[CURPR->T - 1];
       il->MPITYPE = il->S[CURPR->T - 2];
       il->MPICNT = il->S[CURPR->T - 3];
@@ -1065,7 +1065,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
           }
         }
       }
-      for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; xl.I++) {
+      for (xl.I = 0; xl.I <= HighestProcessor; xl.I++) {
         for (xl.J = 0; xl.J < il->MPICNT; xl.J++) {
           il->S[il->MPIPNT[xl.I] + xl.J] = il->S[xl.H1 + xl.J];
           if (il->MPITYPE == MPIFLOAT) {
@@ -1101,7 +1101,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       break;
     }
     if (CHKCOMM(&xl, il->S[CURPR->T]) || (il->S[CURPR->T - 1] < 0) ||
-        (il->S[CURPR->T - 1] > HIGHESTPROCESSOR) ||
+        (il->S[CURPR->T - 1] > HighestProcessor) ||
         ((il->S[CURPR->T - 2] != MPIINT && il->S[CURPR->T - 2] != MPIFLOAT &&
           il->S[CURPR->T - 2] != MPICHAR)) ||
         (il->S[CURPR->T - 3] <= 0) || (il->S[CURPR->T - 4] <= 0) ||
@@ -1116,7 +1116,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 60;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
       il->MPIROOT = il->S[CURPR->T - 1];
       il->MPITYPE = il->S[CURPR->T - 5];
       il->MPICNT = il->S[CURPR->T - 6];
@@ -1154,13 +1154,13 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
               COMMDELAY(il, CURPR->PROCESSOR, il->MPIROOT, il->MPICNT);
     } else {
       xl.R1 = CURPR->TIME +
-              HIGHESTPROCESSOR * (MPINODETIME + std::round(il->MPICNT / 3));
+              HighestProcessor * (MPINODETIME + std::round(il->MPICNT / 3));
     }
     if (il->MPITIME < xl.R1)
       il->MPITIME = xl.R1;
     if (il->MPISEM == 0) {
       xl.H1 = il->MPIRES[il->MPIROOT];
-      for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; ++xl.I) {
+      for (xl.I = 0; xl.I <= HighestProcessor; ++xl.I) {
         for (xl.J = 0; xl.J <= il->MPICNT - 1; ++xl.J) {
           il->S[xl.H1 + il->MPICNT * xl.I + xl.J] =
               il->S[il->MPIPNT[xl.I] + xl.J];
@@ -1193,7 +1193,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       break;
     }
     if (CHKCOMM(&xl, il->S[CURPR->T]) || (il->S[CURPR->T - 1] < 0) ||
-        (il->S[CURPR->T - 1] > HIGHESTPROCESSOR) ||
+        (il->S[CURPR->T - 1] > HighestProcessor) ||
         (!(il->S[CURPR->T - 2] == MPIINT || il->S[CURPR->T - 2] == MPIFLOAT ||
            il->S[CURPR->T - 2] == MPICHAR)) ||
         (il->S[CURPR->T - 3] <= 0) || (il->S[CURPR->T - 4] <= 0) ||
@@ -1206,7 +1206,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 60;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
       il->MPIROOT = il->S[CURPR->T - 1];
       il->MPITYPE = il->S[CURPR->T - 2];
       il->MPICNT = il->S[CURPR->T - 3];
@@ -1248,7 +1248,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     }
     if (il->MPISEM == 0) {
       xl.H1 = il->MPIPNT[il->MPIROOT];
-      for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; xl.I++) {
+      for (xl.I = 0; xl.I <= HighestProcessor; xl.I++) {
         for (xl.J = 0; xl.J < il->MPICNT; xl.J++) {
           il->S[il->MPIRES[xl.I] + xl.J] =
               il->S[xl.H1 + il->MPICNT * xl.I + xl.J];
@@ -1266,7 +1266,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
           proc->STATE = PROCESSDESCRIPTOR::STATE::DELAYED;
           proc->WAKETIME =
               il->MPITIME +
-              (HIGHESTPROCESSOR * (MPINODETIME + (int)(il->MPICNT / 3)));
+              (HighestProcessor * (MPINODETIME + (int)(il->MPICNT / 3)));
         } else {
           xl.R1 = il->MPITIME +
                   COMMDELAY(il, il->MPIROOT, proc->PROCESSOR, il->MPICNT);
@@ -1294,7 +1294,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       break;
     }
     if (CHKCOMM(&xl, il->S[CURPR->T]) || (il->S[CURPR->T - 1] < 0) ||
-        (il->S[CURPR->T - 1] > HIGHESTPROCESSOR) ||
+        (il->S[CURPR->T - 1] > HighestProcessor) ||
         (!(il->S[CURPR->T - 2] == MPIMAX || il->S[CURPR->T - 2] == MPIMIN ||
            il->S[CURPR->T - 2] == MPISUM || il->S[CURPR->T - 2] == MPIPROD ||
            il->S[CURPR->T - 2] == MPILAND || il->S[CURPR->T - 2] == MPILOR ||
@@ -1311,7 +1311,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 61;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
       il->MPIOP = il->S[CURPR->T - 2];
       il->MPITYPE = il->S[CURPR->T - 3];
       il->MPICNT = il->S[CURPR->T - 4];
@@ -1356,7 +1356,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
           } else {
             xl.H2 = 1;
           }
-          for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; xl.I++) {
+          for (xl.I = 0; xl.I <= HighestProcessor; xl.I++) {
             xl.H3 = il->S[il->MPIPNT[xl.I] + xl.J];
             switch (il->MPIOP) {
             case MPIMAX:
@@ -1416,7 +1416,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
           if (il->MPIOP == MPIPROD) {
             xl.R2 = 1.0;
           }
-          for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; xl.I++) {
+          for (xl.I = 0; xl.I <= HighestProcessor; xl.I++) {
             if (il->S[il->MPIPNT[xl.I] + xl.J] == RTAG) {
               xl.R3 = il->RS[il->MPIPNT[xl.I] + xl.J];
             } else if (il->S[il->MPIPNT[xl.I] + xl.J] == 0) {
@@ -1492,7 +1492,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 62;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
       il->MPIOP = il->S[CURPR->T - 1];
       il->MPITYPE = il->S[CURPR->T - 2];
       il->MPICNT = il->S[CURPR->T - 3];
@@ -1513,7 +1513,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       break;
     }
     il->MPISEM = il->MPISEM - 1;
-    xl.H1 = HIGHESTPROCESSOR / 2;
+    xl.H1 = HighestProcessor / 2;
     if (CURPR->PID != xl.H1) {
       xl.R1 = CURPR->TIME + GRPDELAY(il, CURPR->PROCESSOR, xl.H1, il->MPICNT);
     } else {
@@ -1533,7 +1533,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
           } else {
             xl.H2 = 1;
           }
-          for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; xl.I++) {
+          for (xl.I = 0; xl.I <= HighestProcessor; xl.I++) {
             xl.H3 = il->S[il->MPIPNT[xl.I] + xl.J];
             switch (il->MPIOP) {
             case MPIMAX:
@@ -1593,7 +1593,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
           if (il->MPIOP == MPIPROD) {
             xl.R2 = 1.0;
           }
-          for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; xl.I++) {
+          for (xl.I = 0; xl.I <= HighestProcessor; xl.I++) {
             if (il->S[il->MPIPNT[xl.I] + xl.J] == RTAG) {
               xl.R3 = il->RS[il->MPIPNT[xl.I] + xl.J];
             } else if (il->S[il->MPIPNT[xl.I] + xl.J] == 0) {
@@ -1625,7 +1625,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
           il->RS[il->MPIRES[xl.H1] + xl.J] = xl.R2;
         }
       }
-      for (xl.I = 0; xl.I <= HIGHESTPROCESSOR; ++xl.I) {
+      for (xl.I = 0; xl.I <= HighestProcessor; ++xl.I) {
         for (xl.J = 0; xl.J < il->MPICNT; ++xl.J) {
           if (il->MPITYPE == MPIINT) {
             il->S[il->MPIRES[xl.I] + xl.J] = il->S[il->MPIRES[xl.H1] + xl.J];
@@ -1667,7 +1667,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
         (il->S[CURPR->T - 2] != MPICOMMWORLD) ||
         ((il->S[CURPR->T - 3] < 0) && (il->S[CURPR->T - 3] != MPIANYTAG)) ||
         ((il->S[CURPR->T - 4] < 0) && (il->S[CURPR->T - 4] != MPIANYSOURCE)) ||
-        (il->S[CURPR->T - 4] > HIGHESTPROCESSOR)) {
+        (il->S[CURPR->T - 4] > HighestProcessor)) {
       il->PS = InterpLocal::PS::MPIPARCHK;
       break;
     }
@@ -1734,7 +1734,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 64;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
       xl.H1 = -1;
       for (xl.I = CARTMAX; xl.I >= 1; xl.I--) {
         if (il->MPICART[xl.I][0] == -1) {
@@ -1758,7 +1758,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
         il->MPIPER[xl.H1][xl.I] = xl.B1;
         xl.H4 = xl.H4 * il->S[xl.H3 + xl.I - 1];
       }
-      if (xl.H4 != HIGHESTPROCESSOR + 1) {
+      if (xl.H4 != HighestProcessor + 1) {
         il->PS = InterpLocal::PS::MPIPARCHK;
         break;
       }
@@ -1792,7 +1792,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     }
     il->S[il->S[CURPR->T]] = il->MPICOMM;
     il->MPISEM = il->MPISEM - 1;
-    xl.H1 = HIGHESTPROCESSOR / 2;
+    xl.H1 = HighestProcessor / 2;
     if (CURPR->PID != xl.H1) {
       xl.R1 = CURPR->TIME + GRPDELAY(il, CURPR->PROCESSOR, xl.H1, 1);
     } else {
@@ -1923,7 +1923,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->S[CURPR->T - 3] <= CARTSTART ||
         il->S[CURPR->T - 3] > CARTSTART + CARTMAX ||
         il->S[CURPR->T - 1] > MAXDIM || il->S[CURPR->T - 2] < 0 ||
-        il->S[CURPR->T - 2] > HIGHESTPROCESSOR) {
+        il->S[CURPR->T - 2] > HighestProcessor) {
       il->PS = InterpLocal::PS::MPIPARCHK;
       break;
     }
@@ -2002,7 +2002,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
     if (il->MPICODE == -1) {
       il->MPICODE = 70;
       il->MPITIME = 0;
-      il->MPISEM = HIGHESTPROCESSOR + 1;
+      il->MPISEM = HighestProcessor + 1;
       il->MPICNT = il->S[xl.H1];
       if (il->MPICART[il->MPICNT - CARTSTART][0] == -1) {
         il->PS = InterpLocal::PS::MPIPARCHK;
@@ -2019,7 +2019,7 @@ void EXECLIB(InterpLocal *il, ExLocal *el, PROCPNT CURPR, int lID) {
       }
     }
     il->MPISEM = il->MPISEM - 1;
-    xl.H1 = HIGHESTPROCESSOR / 2;
+    xl.H1 = HighestProcessor / 2;
     if (CURPR->PID != xl.H1) {
       xl.R1 = CURPR->TIME + GRPDELAY(il, CURPR->PROCESSOR, xl.H1, 1);
     } else {
