@@ -11,124 +11,130 @@ extern void NEXTCH();
 extern void VARIABLEDECLARATION(BlockLocal *);
 extern void CONSTANTDECLARATION(BlockLocal *);
 extern void TYPEDECLARATION(BlockLocal *);
-extern long LOC(BlockLocal *, ALFA ID);
-extern void SKIP(SYMSET, int);
-extern void TEST(SYMSET &, SYMSET &, int N);
-extern void CALL(BlockLocal *, SYMSET &, int);
+extern auto LOC(BlockLocal *, ALFA ID) -> int64_t;
+extern void SKIP(SymbolSet, int);
+extern void TEST(SymbolSet &, SymbolSet &, int N);
+extern void CALL(BlockLocal *, SymbolSet &, int);
 extern void STANDPROC(BlockLocal *, int);
 extern void ENTER(BlockLocal *, ALFA, OBJECTS);
 extern void ENTERBLOCK();
-extern bool TYPE_COMPATIBLE(ITEM X, ITEM Y);
-extern void BASICEXPRESSION(BlockLocal *, SYMSET, ITEM &);
-extern void CONSTANT(BlockLocal *, SYMSET &, CONREC &);
-void EXPRESSION(BlockLocal *, SYMSET FSYS, ITEM &X);
-void COMPOUNDSTATEMENT(BlockLocal *);
-void FIXBREAKS(long LC1);
-void FIXCONTS(long LC1);
-void SWITCHSTATEMENT(BlockLocal *);
-void IFSTATEMENT(BlockLocal *);
-void DOSTATEMENT(BlockLocal *);
-void WHILESTATEMENT(BlockLocal *);
-void FORSTATEMENT(BlockLocal *);
-void BLOCKSTATEMENT(BlockLocal *);
-void FORALLSTATEMENT(BlockLocal *);
-void FORKSTATEMENT(BlockLocal *);
-void JOINSTATEMENT(BlockLocal *);
-void RETURNSTATEMENT(BlockLocal *);
-void BREAKSTATEMENT(BlockLocal *);
-void CONTINUESTATEMENT();
-void MOVESTATEMENT(BlockLocal *);
-void INPUTSTATEMENT(BlockLocal *);
-void COUTMETHODS(BlockLocal *);
-void OUTPUTSTATEMENT(BlockLocal *);
-static int CONTPNT; // range 0..LOOPMAX
-static long CONTLOC[LOOPMAX + 1];
+extern auto TYPE_COMPATIBLE(Item X, Item Y) -> bool;
+extern void BASICEXPRESSION(BlockLocal *, SymbolSet, Item &);
+extern void CONSTANT(BlockLocal *, SymbolSet &, CONREC &);
 
-void STATEMENT(BlockLocal *bl, SYMSET &FSYS) {
-  long int I;
-  ITEM X;
-  long int JUMPLC;
-  SYMBOL SYMSAV;
-  SYMSET su, sv;
-  SYMSAV = SY;
+void EXPRESSION(BlockLocal *, SymbolSet FSYS, Item &X);
+void COMPOUNDSTATEMENT(BlockLocal * /*bl*/);
+void FIXBREAKS(int64_t LC1);
+void FIXCONTS(int64_t LC1);
+void SWITCHSTATEMENT(BlockLocal * /*block_local*/);
+void IFSTATEMENT(BlockLocal * /*bl*/);
+void DOSTATEMENT(BlockLocal * /*bl*/);
+void WHILESTATEMENT(BlockLocal * /*bl*/);
+void FORSTATEMENT(BlockLocal * /*bl*/);
+void BLOCKSTATEMENT(BlockLocal * /*bl*/);
+void FORALLSTATEMENT(BlockLocal * /*bl*/);
+void FORKSTATEMENT(BlockLocal * /*bl*/);
+void JOINSTATEMENT(BlockLocal * /*bl*/);
+void RETURNSTATEMENT(BlockLocal * /*bl*/);
+void BREAKSTATEMENT(BlockLocal * /*bl*/);
+void CONTINUESTATEMENT();
+void MOVESTATEMENT(BlockLocal * /*bl*/);
+void INPUTSTATEMENT(BlockLocal * /*bl*/);
+void COUTMETHODS(BlockLocal * /*bl*/);
+void OUTPUTSTATEMENT(BlockLocal * /*bl*/);
+
+static int CONTPNT; // range 0..LOOPMAX
+static std::array<int64_t, LOOPMAX + 1> CONTLOC;
+
+void STATEMENT(BlockLocal *block_local, SymbolSet &FSYS) {
+  int64_t I = 0;
+  Item X{};
+  int64_t JUMPLC = 0;
+  Symbol SYMSAV;
+  SymbolSet su;
+  SymbolSet sv;
+  SYMSAV = symbol;
   // INSYMBOL();
   if (STATBEGSYS[SY] || ASSIGNBEGSYS[SY]) {
-    if (SY == LSETBRACK) // BEGINSY ??
+    if (symbol == Symbol::LSETBRACK) // BEGINSY ??
     {
-      COMPOUNDSTATEMENT(bl);
-    } else if (SY == SWITCHSY) {
-      SWITCHSTATEMENT(bl);
-    } else if (SY == IFSY) {
-      IFSTATEMENT(bl);
-    } else if (SY == DOSY) {
-      DOSTATEMENT(bl);
-    } else if (SY == WHILESY) {
-      WHILESTATEMENT(bl);
-    } else if (SY == FORSY) {
-      FORSTATEMENT(bl);
-    } else if (SY == FORALLSY) {
-      FORALLSTATEMENT(bl);
-    } else if (SY == FORKSY) {
-      FORKSTATEMENT(bl);
-    } else if (SY == JOINSY) {
-      JOINSTATEMENT(bl);
-    } else if (SY == RETURNSY) {
-      RETURNSTATEMENT(bl);
-    } else if (SY == BREAKSY) {
-      BREAKSTATEMENT(bl);
-    } else if (SY == CONTSY) {
+      COMPOUNDSTATEMENT(block_local);
+    } else if (symbol == Symbol::SWITCHSY) {
+      SWITCHSTATEMENT(block_local);
+    } else if (symbol == Symbol::IFSY) {
+      IFSTATEMENT(block_local);
+    } else if (symbol == Symbol::DOSY) {
+      DOSTATEMENT(block_local);
+    } else if (symbol == Symbol::WHILESY) {
+      WHILESTATEMENT(block_local);
+    } else if (symbol == Symbol::FORSY) {
+      FORSTATEMENT(block_local);
+    } else if (symbol == Symbol::FORALLSY) {
+      FORALLSTATEMENT(block_local);
+    } else if (symbol == Symbol::FORKSY) {
+      FORKSTATEMENT(block_local);
+    } else if (symbol == Symbol::JOINSY) {
+      JOINSTATEMENT(block_local);
+    } else if (symbol == Symbol::RETURNSY) {
+      RETURNSTATEMENT(block_local);
+    } else if (symbol == Symbol::BREAKSY) {
+      BREAKSTATEMENT(block_local);
+    } else if (symbol == Symbol::CONTSY) {
       CONTINUESTATEMENT();
-    } else if (SY == MOVESY) {
-      MOVESTATEMENT(bl);
-    } else if (SY == CINSY) {
-      INPUTSTATEMENT(bl);
-    } else if (SY == COUTSY) {
-      OUTPUTSTATEMENT(bl);
-    } else if (SY == IDENT) {
-      I = LOC(bl, ID);
+    } else if (symbol == Symbol::MOVESY) {
+      MOVESTATEMENT(block_local);
+    } else if (symbol == Symbol::CINSY) {
+      INPUTSTATEMENT(block_local);
+    } else if (symbol == Symbol::COUTSY) {
+      OUTPUTSTATEMENT(block_local);
+    } else if (symbol == Symbol::IDENT) {
+      I = LOC(block_local, ID);
       if (I != 0) {
-        if (TAB[I].OBJ == PROZEDURE) {
+        if (TAB[I].object == OBJECTS::PROZEDURE) {
           INSYMBOL();
           if (TAB[I].LEV != 0) {
-            CALL(bl, FSYS, I);
+            CALL(block_local, FSYS, I);
           } else {
-            STANDPROC(bl, TAB[I].ADR);
+            STANDPROC(block_local, TAB[I].address);
           }
         } else {
-          EXPRESSION(bl, FSYS, X);
-          if (X.SIZE > 1)
-            EMIT1(112, X.SIZE);
+          EXPRESSION(block_local, FSYS, X);
+          if (X.size > 1)
+            EMIT1(112, X.size);
           EMIT(111);
         }
       } else
         INSYMBOL();
-    } else if (SY == LPARENT || SY == INCREMENT || SY == DECREMENT ||
-               SY == TIMES) {
-      EXPRESSION(bl, FSYS, X);
+    } else if (SY == Symbol::LPARENT || SY == Symbol::INCREMENT ||
+               SY == Symbol::DECREMENT || SY == Symbol::TIMES) {
+      EXPRESSION(block_local, FSYS, X);
       EMIT(111);
     }
   }
   if (ASSIGNBEGSYS[SYMSAV] ||
-      (SYMSAV == DOSY || SYMSAV == JOINSY || SYMSAV == RETURNSY ||
-       SYMSAV == BREAKSY || SYMSAV == CONTSY || SYMSAV == SEMICOLON ||
-       SYMSAV == MOVESY || SYMSAV == CINSY || SYMSAV == COUTSY ||
-       SYMSAV == RSETBRACK)) // RSETBRACK added for while missing statement DDE
+      (SYMSAV == SYMBOL::DOSY || SYMSAV == SYMBOL::JOINSY ||
+       SYMSAV == SYMBOL::RETURNSY || SYMSAV == SYMBOL::BREAKSY ||
+       SYMSAV == SYMBOL::CONTSY || SYMSAV == SYMBOL::SEMICOLON ||
+       SYMSAV == SYMBOL::MOVESY || SYMSAV == SYMBOL::CINSY ||
+       SYMSAV == SYMBOL::COUTSY ||
+       SYMSAV == SYMBOL::RSETBRACK)) // RSETBRACK added for while missing
+                                     // statement DDE
   {
-    if (SY == SEMICOLON)
+    if (SY == Symbol::SEMICOLON)
       INSYMBOL();
     else
-      ERROR(14);
+      error(14);
   }
   su = 0;
   sv = FSYS;
-  sv[ELSESY] = true; // added to fix else DDE
+  sv[symbol::ELSESY] = true; // added to fix else DDE
   // TEST(FSYS, su, 14);
   TEST(sv, su, 14);
 }
 
 void COMPOUNDSTATEMENT(BlockLocal *bl) {
-  long int X;
-  SYMSET su, sv;
+  int64_t X;
+  SymbolSet su, sv;
   INSYMBOL();
   su = DECLBEGSYS | STATBEGSYS | ASSIGNBEGSYS;
   // su[RSETBRACK] = true;  ??
@@ -136,24 +142,24 @@ void COMPOUNDSTATEMENT(BlockLocal *bl) {
   sv[RSETBRACK] = true;
   TEST(su, sv, 123);
   while (DECLBEGSYS[SY] || STATBEGSYS[SY] || ASSIGNBEGSYS[SY]) {
-    if (SY == DEFINESY) {
+    if (SY == SYMBOL::DEFINESY) {
       CONSTANTDECLARATION(bl);
     }
-    if (SY == TYPESY) {
+    if (SY == SYMBOL::TYPESY) {
       TYPEDECLARATION(bl);
     }
-    if (SY == STRUCTSY) {
+    if (SY == SYMBOL::STRUCTSY) {
       VARIABLEDECLARATION(bl);
     }
-    if (SY == IDENT) {
+    if (SY == SYMBOL::IDENT) {
       X = LOC(bl, ID);
       if (X != 0) {
-        if (TAB[X].OBJ == TYPE1) {
+        if (TAB[X].OBJ == OBJECTS::TYPE1) {
           VARIABLEDECLARATION(bl);
         } else {
           su = bl->FSYS;
-          su[RSETBRACK] = true;
-          su[SEMICOLON] = true;
+          su[SYMBOL::RSETBRACK] = true;
+          su[SYMBOL::SEMICOLON] = true;
           STATEMENT(bl, su);
         }
       } else {
@@ -161,315 +167,319 @@ void COMPOUNDSTATEMENT(BlockLocal *bl) {
       }
     } else if (STATBEGSYS[SY] || ASSIGNBEGSYS[SY]) {
       su = bl->FSYS;
-      su[RSETBRACK] = true;
-      su[SEMICOLON] = true;
+      su[SYMBOL::RSETBRACK] = true;
+      su[SYMBOL::SEMICOLON] = true;
       STATEMENT(bl, su);
     }
     su = DECLBEGSYS | STATBEGSYS | ASSIGNBEGSYS;
-    su[RSETBRACK] = true;
+    su[SYMBOL::RSETBRACK] = true;
     TEST(su, bl->FSYS, 6);
   }
-  if (SYMCNT == 1) {
-    LOCATION[LNUM] = LC;
+  if (symbol_count == 1) {
+    LOCATION[LNUM] = line_count;
   }
-  if (SY == RSETBRACK) {
+  if (SY == SYMBOL::RSETBRACK) {
     INSYMBOL();
   } else {
-    ERROR(104);
+    error(104);
   }
 }
 
-void FIXBREAKS(long LC1) {
-  long LC2;
+void FIXBREAKS(int64_t LC1) {
+  int64_t LC2;
   while (BREAKLOC[BREAKPNT] > 0) {
-    LC2 = CODE[BREAKLOC[BREAKPNT]].Y;
-    CODE[BREAKLOC[BREAKPNT]].Y = LC1;
+    LC2 = code[BREAKLOC[BREAKPNT]].Y;
+    code[BREAKLOC[BREAKPNT]].Y = LC1;
     BREAKLOC[BREAKPNT] = LC2;
   }
   BREAKPNT = BREAKPNT - 1;
 }
 
-void FIXCONTS(long int LC1) {
-  long int LC2;
+void FIXCONTS(int64_t LC1) {
+  int64_t LC2;
   while (CONTLOC[CONTPNT] > 0) {
-    LC2 = CODE[CONTLOC[CONTPNT]].Y;
-    CODE[CONTLOC[CONTPNT]].Y = LC1;
+    LC2 = code[CONTLOC[CONTPNT]].Y;
+    code[CONTLOC[CONTPNT]].Y = LC1;
     CONTLOC[CONTPNT] = LC2;
   }
   CONTPNT = CONTPNT - 1;
 }
 
 void IFSTATEMENT(BlockLocal *bl) {
-  ITEM X = {NOTYP, 0};
+  Item X = {Types::NOTYP, 0};
   int LC1, LC2;
-  SYMSET su;
+  SymbolSet su;
   INSYMBOL();
-  if (SY == LPARENT) {
+  if (SY == Symbol::LPARENT) {
     INSYMBOL();
     su = bl->FSYS;
-    su[RPARENT] = true;
+    su[symbol::RPARENT] = true;
     EXPRESSION(bl, su, X);
-    if (!(X.TYP == BOOLS || X.TYP == INTS || X.TYP == NOTYP)) {
-      ERROR(17);
+    if (!(X.types == Types::BOOLS || X.types == Types::INTS ||
+          X.types == Types::NOTYP)) {
+      error(17);
     }
-    LC1 = LC;
+    LC1 = line_count;
     EMIT(11);
-    if (LOCATION[LNUM] == LC - 1) {
-      LOCATION[LNUM] = LC;
+    if (LOCATION[LNUM] == line_count - 1) {
+      LOCATION[LNUM] = line_count;
     }
-    if (SY == RPARENT) {
+    if (SY == Symbol::RPARENT) {
       INSYMBOL();
     } else {
-      ERROR(4);
+      error(4);
     }
     su = bl->FSYS;
-    su[ELSESY] = true;
+    su[symbol::ELSESY] = true;
     STATEMENT(bl, su);
-    if (SY == ELSESY) {
-      LC2 = LC;
+    if (SY == Symbol::ELSESY) {
+      LC2 = line_count;
       EMIT(10);
-      if (SYMCNT == 1) {
-        LOCATION[LNUM] = LC;
+      if (symbol_count == 1) {
+        LOCATION[LNUM] = line_count;
       }
-      CODE[LC1].Y = LC;
+      code[LC1].Y = line_count;
       INSYMBOL();
       STATEMENT(bl, bl->FSYS);
-      CODE[LC2].Y = LC;
+      code[LC2].Y = line_count;
     } else {
-      CODE[LC1].Y = LC;
+      code[LC1].Y = line_count;
     }
   } else {
-    ERROR(9);
+    error(9);
   }
 }
 
 void DOSTATEMENT(BlockLocal *bl) {
-  ITEM X;
+  Item X;
   int LC1, LC2, LC3;
-  LC1 = LC;
-  SYMSET su;
+  LC1 = line_count;
+  SymbolSet su;
   INSYMBOL();
   BREAKPNT = BREAKPNT + 1;
   BREAKLOC[BREAKPNT] = 0;
   CONTPNT = CONTPNT + 1;
   CONTLOC[CONTPNT] = 0;
   su = bl->FSYS;
-  su[SEMICOLON] = true;
-  su[WHILESY] = true;
+  su[symbol::SEMICOLON] = true;
+  su[symbol::WHILESY] = true;
   STATEMENT(bl, su);
-  LC2 = LC;
-  if (SY == WHILESY) {
-    if (SYMCNT == 1) {
-      LOCATION[LNUM] = LC;
+  LC2 = line_count;
+  if (SY == Symbol::WHILESY) {
+    if (symbol_count == 1) {
+      LOCATION[LNUM] = line_count;
     }
     INSYMBOL();
-    if (SY == LPARENT) {
+    if (SY == Symbol::LPARENT) {
       INSYMBOL();
     } else {
-      ERROR(9);
+      error(9);
     }
     su = bl->FSYS;
-    su[RPARENT] = true;
+    su[symbol::RPARENT] = true;
     EXPRESSION(bl, su, X);
-    if (!(X.TYP == BOOLS || X.TYP == INTS || X.TYP == NOTYP)) {
-      ERROR(17);
+    if (!(X.types == Types::BOOLS || X.types == Types::INTS ||
+          X.types == Types::NOTYP)) {
+      error(17);
     }
-    if (SY == RPARENT) {
+    if (SY == Symbol::RPARENT) {
       INSYMBOL();
     } else {
-      ERROR(4);
+      error(4);
     }
-    EMIT1(11, LC + 2);
+    EMIT1(11, line_count + 2);
     EMIT1(10, LC1);
   } else {
-    ERROR(53);
+    error(53);
   }
-  FIXBREAKS(LC);
+  FIXBREAKS(line_count);
   FIXCONTS(LC2);
 }
 
 void WHILESTATEMENT(BlockLocal *bl) {
-  ITEM X;
+  Item X;
   int LC1, LC2, LC3;
-  SYMSET su;
+  SymbolSet su;
   INSYMBOL();
-  if (SY == LPARENT) {
+  if (SY == Symbol::LPARENT) {
     INSYMBOL();
   } else {
-    ERROR(9);
+    error(9);
   }
-  LC1 = LC;
+  LC1 = line_count;
   BREAKPNT = BREAKPNT + 1;
   BREAKLOC[BREAKPNT] = 0;
   CONTPNT = CONTPNT + 1;
   CONTLOC[CONTPNT] = 0;
   su = bl->FSYS;
-  su[RPARENT] = true;
+  su[symbol::RPARENT] = true;
   EXPRESSION(bl, su, X);
-  if (!(X.TYP == BOOLS || X.TYP == INTS || X.TYP == NOTYP)) {
-    ERROR(17);
+  if (!(X.types == Types::BOOLS || X.types == Types::INTS ||
+        X.types == Types::NOTYP)) {
+    error(17);
   }
-  LC2 = LC;
+  LC2 = line_count;
   EMIT(11);
-  if (SY == RPARENT) {
+  if (SY == Symbol::RPARENT) {
     INSYMBOL();
   } else {
-    ERROR(4);
+    error(4);
   }
   STATEMENT(bl, bl->FSYS);
   EMIT1(10, LC1);
-  CODE[LC2].Y = LC;
-  if (SYMCNT == 1) {
-    LOCATION[LNUM] = LC;
+  code[LC2].Y = line_count;
+  if (symbol_count == 1) {
+    LOCATION[LNUM] = line_count;
   }
-  FIXBREAKS(LC);
-  FIXCONTS(LC - 1);
+  FIXBREAKS(line_count);
+  FIXCONTS(line_count - 1);
 }
 
-void COMMAEXPR(BlockLocal *bl, SYMBOL TESTSY) {
-  ITEM X;
-  SYMSET su;
+void COMMAEXPR(BlockLocal *bl, Symbol TESTSY) {
+  Item X;
+  SymbolSet su;
   su = bl->FSYS;
-  su[COMMA] = true;
+  su[symbol::COMMA] = true;
   su[TESTSY] = true;
   EXPRESSION(bl, su, X);
-  if (X.SIZE > 1) {
-    EMIT1(112, X.SIZE);
+  if (X.Size > 1) {
+    EMIT1(112, X.Size);
   }
   EMIT(111);
-  while (SY == COMMA) {
+  while (SY == Symbol::COMMA) {
     INSYMBOL();
     EXPRESSION(bl, su, X);
-    if (X.SIZE > 1) {
-      EMIT1(112, X.SIZE);
+    if (X.Size > 1) {
+      EMIT1(112, X.Size);
     }
     EMIT(111);
   }
 }
 
 void FORSTATEMENT(BlockLocal *bl) {
-  ITEM X;
+  Item X;
   int LC1, LC2, LC3;
-  SYMSET su;
+  SymbolSet su;
   INSYMBOL();
-  if (SY == LPARENT) {
+  if (SY == Symbol::LPARENT) {
     INSYMBOL();
   } else {
-    ERROR(9);
+    error(9);
   }
   BREAKPNT = BREAKPNT + 1;
   BREAKLOC[BREAKPNT] = 0;
   CONTPNT = CONTPNT + 1;
   CONTLOC[CONTPNT] = 0;
-  if (SY != SEMICOLON) {
-    COMMAEXPR(bl, SEMICOLON);
+  if (SY != Symbol::SEMICOLON) {
+    COMMAEXPR(bl, Symbol::SEMICOLON);
   }
-  if (SY == SEMICOLON) {
+  if (SY == Symbol::SEMICOLON) {
     INSYMBOL();
   } else {
-    ERROR(14);
+    error(14);
   }
-  LC1 = LC;
-  if (SY != SEMICOLON) {
+  LC1 = line_count;
+  if (SY != Symbol::SEMICOLON) {
     su = bl->FSYS;
-    su[SEMICOLON] = true;
+    su[symbol::SEMICOLON] = true;
     EXPRESSION(bl, su, X);
-    if (!(X.TYP == BOOLS || X.TYP == INTS || X.TYP == NOTYP)) {
-      ERROR(17);
+    if (!(X.types == Types::BOOLS || X.types == Types::INTS ||
+          X.types == Types::NOTYP)) {
+      error(17);
     }
   } else {
     EMIT1(24, 1);
   }
-  if (SY == SEMICOLON) {
+  if (SY == Symbol::SEMICOLON) {
     INSYMBOL();
   } else {
-    ERROR(14);
+    error(14);
   }
-  LC2 = LC;
+  LC2 = line_count;
   EMIT(11);
   EMIT(10);
-  LC3 = LC;
-  if (SY != RPARENT) {
-    COMMAEXPR(bl, RPARENT);
+  LC3 = line_count;
+  if (SY != Symbol::RPARENT) {
+    COMMAEXPR(bl, Symbol::RPARENT);
   }
   EMIT1(10, LC1);
-  if (SY == RPARENT) {
+  if (SY == Symbol::RPARENT) {
     INSYMBOL();
   } else {
-    ERROR(4);
+    error(4);
   }
-  CODE[LC2 + 1].Y = LC;
+  code[LC2 + 1].Y = line_count;
   STATEMENT(bl, bl->FSYS);
   EMIT1(10, LC3);
-  CODE[LC2].Y = LC;
-  if (SYMCNT == 1) {
-    LOCATION[LNUM] = LC;
+  code[LC2].Y = line_count;
+  if (symbol_count == 1) {
+    LOCATION[LNUM] = line_count;
   }
-  FIXBREAKS(LC);
-  FIXCONTS(LC - 1);
+  FIXBREAKS(line_count);
+  FIXCONTS(line_count - 1);
 }
 
 void BLOCKSTATEMENT(BlockLocal *bl) {
-  TYPES CVT;
-  ITEM X;
-  SYMSET su, sv;
-  long int SAVEDX, SAVENUMWITH, SAVEMAXNUMWITH, PRB, PRT, I;
+  Types CVT;
+  Item X;
+  SymbolSet su, sv;
+  int64_t SAVEDX, SAVENUMWITH, SAVEMAXNUMWITH, PRB, PRT, I;
   bool TCRFLAG;
   // INSYMBOL();
-  strcpy(ID, DUMMYNAME);
-  ENTER(bl, ID, PROZEDURE);
-  DUMMYNAME[13] = (char)(DUMMYNAME[13] + 1);
-  if (DUMMYNAME[13] == '0') {
-    DUMMYNAME[12] = (char)(DUMMYNAME[12] + 1);
+  ID = dummy_name;
+  ENTER(bl, ID, OBJECTS::PROZEDURE);
+  dummy_name[13] = static_cast<char>(dummy_name[13] + 1);
+  if (dummy_name[13] == '0') {
+    dummy_name[12] = static_cast<char>(dummy_name[12] + 1);
   }
-  TAB[Tx].NORMAL = true;
+  TAB[tab_index].normal = true;
   SAVEDX = bl->DX;
   SAVENUMWITH = bl->NUMWITH;
   SAVEMAXNUMWITH = bl->MAXNUMWITH;
   bl->LEVEL = bl->LEVEL + 1;
   bl->DX = BASESIZE;
-  PRT = Tx;
+  PRT = tab_index;
   if (bl->LEVEL > LMAX) {
-    FATAL(5);
+    fatal(5);
   }
   bl->NUMWITH = 0;
   bl->MAXNUMWITH = 0;
   ENTERBLOCK();
-  DISPLAY[bl->LEVEL] = B;
-  PRB = B;
-  TAB[PRT].TYP = NOTYP;
-  TAB[PRT].REF = PRB;
-  BTAB[PRB].LASTPAR = Tx;
+  DISPLAY[bl->LEVEL] = btab_index;
+  PRB = btab_index;
+  TAB[PRT].types = Types::NOTYP;
+  TAB[PRT].reference = PRB;
+  BTAB[PRB].LASTPAR = tab_index;
   BTAB[PRB].PSIZE = bl->DX;
   su = 0;
-  su[LSETBRACK] = true;
+  su[symbol::LSETBRACK] = true;
   sv = DECLBEGSYS | STATBEGSYS | ASSIGNBEGSYS;
   TEST(su, sv, 3);
   INSYMBOL();
   EMIT1(18, PRT);
   EMIT1(19, BTAB[PRB].PSIZE - 1);
-  TAB[PRT].ADR = LC;
+  TAB[PRT].ADR = line_count;
   su = bl->FSYS;
-  su[RSETBRACK] = true;
+  su[symbol::RSETBRACK] = true;
   // sv = DECLBEGSYS | STATBEGSYS | ASSIGNBEGSYS;
   TEST(sv, su, 101);
   // while (DECLBEGSYS[SY] || STATBEGSYS[SY] || ASSIGNBEGSYS[SY])
   while (sv[SY]) {
-    if (SY == DEFINESY) {
+    if (SY == Symbol::DEFINESY) {
       CONSTANTDECLARATION(bl);
-    } else if (SY == TYPESY) {
+    } else if (SY == Symbol::TYPESY) {
       TYPEDECLARATION(bl);
-    } else if (SY == STRUCTSY) {
+    } else if (SY == Symbol::STRUCTSY) {
       VARIABLEDECLARATION(bl);
-    } else if (SY == IDENT) {
+    } else if (SY == Symbol::IDENT) {
       I = LOC(bl, ID);
       if (I != 0) {
         if (TAB[I].OBJ == TYPE1) {
           VARIABLEDECLARATION(bl);
         } else {
           su = bl->FSYS;
-          su[SEMICOLON] = true;
-          su[RSETBRACK] = true;
+          su[symbol::SEMICOLON] = true;
+          su[symbol::RSETBRACK] = true;
           STATEMENT(bl, su);
         }
       } else {
@@ -477,26 +487,26 @@ void BLOCKSTATEMENT(BlockLocal *bl) {
       }
     } else if (STATBEGSYS[SY] || ASSIGNBEGSYS[SY]) {
       su = bl->FSYS;
-      su[SEMICOLON] = true;
-      su[RSETBRACK] = true;
+      su[symbol::SEMICOLON] = true;
+      su[symbol::RSETBRACK] = true;
       STATEMENT(bl, su);
     }
     su = DECLBEGSYS | STATBEGSYS | ASSIGNBEGSYS;
-    su[RSETBRACK] = true;
+    su[symbol::RSETBRACK] = true;
     TEST(su, bl->FSYS, 6);
   }
   BTAB[PRB].VSIZE = bl->DX;
   BTAB[PRB].VSIZE = BTAB[PRB].VSIZE + bl->MAXNUMWITH;
-  if (SYMCNT == 1) {
-    LOCATION[LNUM] = LC;
+  if (symbol_count == 1) {
+    LOCATION[LNUM] = line_count;
   }
-  if (SY == RSETBRACK) {
+  if (SY == Symbol::RSETBRACK) {
     INSYMBOL();
   } else {
-    ERROR(7);
+    error(7);
   }
   su = bl->FSYS;
-  su[PERIOD] = true;
+  su[symbol::PERIOD] = true;
   sv = 0;
   TEST(su, sv, 6);
   EMIT(105);
@@ -507,118 +517,119 @@ void BLOCKSTATEMENT(BlockLocal *bl) {
 }
 
 void FORALLSTATEMENT(BlockLocal *bl) {
-  TYPES CVT;
-  ITEM X;
-  long int I, J, LC1, LC2, LC3;
+  Types CVT;
+  Item X;
+  int64_t I, J, LC1, LC2, LC3;
   // bool DEFOUND;
   bool TCRFLAG;
-  SYMSET su;
+  SymbolSet su;
   bl->FLEVEL = bl->FLEVEL + 1;
   INSYMBOL();
-  if (SY == IDENT) {
+  if (SY == Symbol::IDENT) {
     I = LOC(bl, ID);
     TAB[I].FORLEV = bl->FLEVEL;
     INSYMBOL();
     if (I == 0) {
-      CVT = INTS;
+      CVT = Types::INTS;
     } else {
-      if (TAB[I].OBJ == VARIABLE) {
-        CVT = TAB[I].TYP;
-        if (!TAB[I].NORMAL) {
-          ERROR(26);
+      if (TAB[I].object == OBJECTS::VARIABLE) {
+        CVT = TAB[I].types;
+        if (!TAB[I].normal) {
+          error(26);
         } else {
-          EMIT2(0, TAB[I].LEV, TAB[I].ADR);
+          EMIT2(0, TAB[I].LEV, TAB[I].address);
         }
-        if (!(CVT == NOTYP || CVT == INTS || CVT == BOOLS || CVT == CHARS)) {
-          ERROR(18);
+        if (!(CVT == Types::NOTYP || CVT == Types::INTS ||
+              CVT == Types::BOOLS || CVT == Types::CHARS)) {
+          error(18);
         }
       } else {
-        ERROR(37);
-        CVT = INTS;
+        error(37);
+        CVT = Types::INTS;
       }
     }
   } else {
     su = bl->FSYS;
-    su[BECOMES] = true;
-    su[TOSY] = true;
-    su[DOSY] = true;
-    su[GROUPINGSY] = true;
+    su[symbol::BECOMES] = true;
+    su[symbol::TOSY] = true;
+    su[symbol::DOSY] = true;
+    su[symbol::GROUPINGSY] = true;
     SKIP(su, 2);
-    CVT = INTS;
+    CVT = Types::INTS;
   }
-  if (SY == BECOMES) {
+  if (SY == Symbol::BECOMES) {
     INSYMBOL();
     su = bl->FSYS;
-    su[TOSY] = true;
-    su[DOSY] = true;
-    su[GROUPINGSY] = true;
+    su[symbol::TOSY] = true;
+    su[symbol::DOSY] = true;
+    su[symbol::GROUPINGSY] = true;
     EXPRESSION(bl, su, X);
-    if (X.TYP != CVT) {
-      ERROR(19);
+    if (X.types != CVT) {
+      error(19);
     }
   } else {
     su = bl->FSYS;
-    su[TOSY] = true;
-    su[DOSY] = true;
-    su[GROUPINGSY] = true;
+    su[symbol::TOSY] = true;
+    su[symbol::DOSY] = true;
+    su[symbol::GROUPINGSY] = true;
     SKIP(su, 51);
   }
-  if (SY == TOSY) {
+  if (SY == Symbol::TOSY) {
     INSYMBOL();
     su = bl->FSYS;
-    su[ATSY] = true;
-    su[DOSY] = true;
-    su[GROUPINGSY] = true;
+    su[symbol::ATSY] = true;
+    su[symbol::DOSY] = true;
+    su[symbol::GROUPINGSY] = true;
     EXPRESSION(bl, su, X);
-    if (X.TYP != CVT) {
-      ERROR(19);
+    if (X.types != CVT) {
+      error(19);
     }
   } else {
     su = bl->FSYS;
-    su[ATSY] = true;
-    su[DOSY] = true;
-    su[GROUPINGSY] = true;
+    su[symbol::ATSY] = true;
+    su[symbol::DOSY] = true;
+    su[symbol::GROUPINGSY] = true;
     SKIP(su, 55);
   }
-  if (SY == GROUPINGSY) {
+  if (SY == Symbol::GROUPINGSY) {
     INSYMBOL();
     su = bl->FSYS;
-    su[ATSY] = true;
-    su[DOSY] = true;
+    su[symbol::ATSY] = true;
+    su[symbol::DOSY] = true;
     EXPRESSION(bl, su, X);
-    if (X.TYP != INTS) {
-      ERROR(45);
+    if (X.types != Types::INTS) {
+      error(45);
     }
   } else {
     EMIT1(24, 1);
   }
   EMIT(4);
-  LC1 = LC;
+  LC1 = line_count;
   EMIT(75);
-  LC2 = LC;
+  LC2 = line_count;
   TAB[I].FORLEV = -TAB[I].FORLEV;
-  if (SY == ATSY) {
+  if (SY == Symbol::ATSY) {
     INSYMBOL();
     su = bl->FSYS;
-    su[DOSY] = true;
+    su[symbol::DOSY] = true;
     EXPRESSION(bl, su, X);
-    if (!(X.TYP == INTS || X.TYP == NOTYP)) {
-      ERROR(126);
+    if (!(X.types == Types::INTS || X.types == Types::NOTYP)) {
+      error(126);
     }
   } else {
     EMIT(79);
   }
-  if (SY == DOSY) {
+  if (SY == Symbol::DOSY) {
     INSYMBOL();
   } else {
-    ERROR(54);
+    error(54);
   }
   TAB[I].FORLEV = -TAB[I].FORLEV;
   bl->CREATEFLAG = false;
-  if (SY == IDENT) {
+  if (SY == Symbol::IDENT) {
     J = LOC(bl, ID);
     if (J != 0) {
-      if (TAB[J].OBJ == PROZEDURE && TAB[J].LEV != 0) {
+      if (TAB[J].object == OBJECTS::PROZEDURE && TAB[J].LEV != 0) {
         bl->CREATEFLAG = true;
       }
     }
@@ -629,9 +640,9 @@ void FORALLSTATEMENT(BlockLocal *bl) {
     EMIT1(74, 0);
   }
   TCRFLAG = bl->CREATEFLAG;
-  LC3 = LC;
+  LC3 = line_count;
   EMIT1(10, 0);
-  if (SY == LSETBRACK) {
+  if (SY == Symbol::LSETBRACK) {
     BLOCKSTATEMENT(bl);
   } else {
     su = bl->FSYS;
@@ -644,49 +655,49 @@ void FORALLSTATEMENT(BlockLocal *bl) {
     EMIT2(104, LC3 + 1, 0);
   }
   EMIT(70);
-  CODE[LC3].Y = LC;
+  code.at(LC3).Y = line_count;
   EMIT1(76, LC2);
-  CODE[LC1].Y = LC;
+  code.at(LC1).Y = line_count;
   EMIT(5);
   bl->FLEVEL = bl->FLEVEL - 1;
-  if (SYMCNT == 1) {
-    LOCATION[LNUM] = LC;
+  if (symbol_count == 1) {
+    LOCATION[LNUM] = line_count;
   }
 }
 
 void FORKSTATEMENT(BlockLocal *bl) {
-  ITEM X;
+  Item X;
   int I; // added local (unlikely intended to use global)
   int JUMPLC;
-  SYMSET su;
+  SymbolSet su;
   bl->CREATEFLAG = false;
   EMIT(106);
   INSYMBOL();
-  if (SY == LPARENT) {
+  if (SY == Symbol::LPARENT) {
     INSYMBOL();
-    if (SY == ATSY) {
+    if (SY == Symbol::ATSY) {
       INSYMBOL();
       su = bl->FSYS;
-      su[RPARENT] = true;
+      su[symbol::RPARENT] = true;
       EXPRESSION(bl, su, X);
-      if (!(X.TYP == INTS || X.TYP == NOTYP)) {
-        ERROR(126);
+      if (!(X.types == Types::INTS || X.types == Types::NOTYP)) {
+        error(126);
       }
     } else {
-      ERROR(127);
+      error(127);
     }
-    if (SY == RPARENT) {
+    if (SY == Symbol::RPARENT) {
       INSYMBOL();
     } else {
-      ERROR(4);
+      error(4);
     }
   } else {
     EMIT(79);
   }
-  if (SY == IDENT) {
+  if (SY == Symbol::IDENT) {
     I = LOC(bl, ID);
     if (I != 0) {
-      if (TAB[I].OBJ == PROZEDURE && TAB[I].LEV != 0) {
+      if (TAB[I].object == OBJECTS::PROZEDURE && TAB[I].LEV != 0) {
         bl->CREATEFLAG = true;
       }
     }
@@ -696,13 +707,13 @@ void FORKSTATEMENT(BlockLocal *bl) {
   } else {
     EMIT1(67, 0);
   }
-  JUMPLC = LC;
+  JUMPLC = line_count;
   EMIT1(7, 0);
   STATEMENT(bl, bl->FSYS);
   EMIT(69);
-  CODE[JUMPLC].Y = LC;
-  if (SYMCNT == 1) {
-    LOCATION[LNUM] = LC;
+  code.at(JUMPLC).Y = line_count;
+  if (symbol_count == 1) {
+    LOCATION[LNUM] = line_count;
   }
 }
 
@@ -712,17 +723,17 @@ void JOINSTATEMENT(BlockLocal *bl) {
 }
 
 void RETURNSTATEMENT(BlockLocal *bl) {
-  ITEM X;
+  Item X;
   INSYMBOL();
-  if (RETURNTYPE.TYP != VOIDS) {
+  if (RETURNTYPE.types != Types::VOIDS) {
     EXPRESSION(bl, bl->FSYS, X);
-    if (X.TYP == RECS || X.TYP == ARRAYS) {
-      EMIT1(113, X.SIZE);
+    if (X.types == Types::RECS || X.types == Types::ARRAYS) {
+      EMIT1(113, X.size);
     }
-    if (!TYPE_COMPATIBLE(RETURNTYPE, X)) {
-      ERROR(118);
+    if (!TYPE_COMPATIBLE(ReturnType, X)) {
+      error(118);
     } else {
-      if (RETURNTYPE.TYP == REALS && X.TYP == INTS) {
+      if (RETURNTYPE.types == Types::REALS && X.types == Types::INTS) {
         EMIT(26);
       }
       EMIT(33);
@@ -734,33 +745,33 @@ void RETURNSTATEMENT(BlockLocal *bl) {
 
 void BREAKSTATEMENT(BlockLocal *bl) {
   EMIT1(10, BREAKLOC[BREAKPNT]);
-  BREAKLOC[BREAKPNT] = LC - 1;
+  BREAKLOC[BREAKPNT] = line_count - 1;
   INSYMBOL();
 }
 
 void CONTINUESTATEMENT() {
   EMIT1(10, CONTLOC[CONTPNT]);
-  CONTLOC[CONTPNT] = LC - 1;
+  CONTLOC[CONTPNT] = line_count - 1;
   INSYMBOL();
 }
 
 void MOVESTATEMENT(BlockLocal *bl) {
-  ITEM X;
-  SYMSET su;
+  Item X;
+  SymbolSet su;
   INSYMBOL();
   su = bl->FSYS;
-  su[TOSY] = true;
+  su[symbol::TOSY] = true;
   BASICEXPRESSION(bl, su, X);
-  if (!(X.ISADDR && X.TYP == CHANS)) {
-    ERROR(103);
+  if (!(X.ISADDR && X.types == Types::CHANS)) {
+    error(103);
   }
-  if (SY == TOSY) {
+  if (SY == Symbol::TOSY) {
     INSYMBOL();
     su = bl->FSYS;
-    su[SEMICOLON] = true;
+    su[symbol::SEMICOLON] = true;
     EXPRESSION(bl, su, X);
-    if (!(X.TYP == INTS || X.TYP == NOTYP)) {
-      ERROR(126);
+    if (!(X.types == Types::INTS || X.types == Types::NOTYP)) {
+      error(126);
     }
   } else {
     EMIT1(8, 19);
@@ -795,7 +806,7 @@ void MOVESTATEMENT(BlockLocal *bl) {
 //                    EXPRESSION(bl, su, X);
 //                    if (!X.ISADDR)
 //                    {
-//                        ERROR(131);
+//                        error(131);
 //                    } else
 //                    {
 //                        if (X.TYP == INTS || X.TYP == REALS || X.TYP == CHARS
@@ -804,69 +815,69 @@ void MOVESTATEMENT(BlockLocal *bl) {
 //                            EMIT1(27, (int)X.TYP);
 //                        } else
 //                        {
-//                            ERROR(40);
+//                            error(40);
 //                        }
 //                    }
 //                }
 //            } while (SY == INSTR);
 //        } else
 //        {
-//            ERROR(129);
+//            error(129);
 //        }
 //    }
 void INPUTSTATEMENT(BlockLocal *bl) {
-  ITEM X;
-  SYMSET su;
+  Item X;
+  SymbolSet su;
   INSYMBOL();
-  if (SY == INSTR) {
+  if (SY == Symbol::INSTR) {
     do {
       INSYMBOL();
       su = bl->FSYS;
-      su[INSTR] = true;
-      su[SEMICOLON] = true;
+      su[symbol::INSTR] = true;
+      su[symbol::SEMICOLON] = true;
       BASICEXPRESSION(bl, su, X);
-      if (!X.ISADDR) {
-        ERROR(131);
+      if (!X.is_address) {
+        error(131);
       } else {
-        if (X.TYP == INTS || X.TYP == REALS || X.TYP == CHARS ||
-            X.TYP == NOTYP) {
-          EMIT1(27, (int)X.TYP);
+        if (X.types == Types::INTS || X.types == Types::REALS ||
+            X.types == Types::CHARS || X.types == Types::NOTYP) {
+          EMIT1(27, (int)X.types);
         } else {
-          ERROR(40);
+          error(40);
         }
       }
-    } while (SY == INSTR);
+    } while (SY == Symbol::INSTR);
   } else {
-    ERROR(129);
+    error(129);
   }
 }
 
 void COUTMETHODS(BlockLocal *bl) {
   ALFA METHOD;
-  ITEM X;
-  SYMSET su;
+  Item X;
+  SymbolSet su;
   INSYMBOL();
-  strcpy(METHOD, ID);
-  if (SY != IDENT || (strcmp(ID, "WIDTH         ") != 0 &&
-                      strcmp(ID, "PRECISION     ") != 0)) {
-    ERROR(133);
+  METHOD = ID;
+  if (SY != Symbol::IDENT || (strcmp(ID.c_str(), "WIDTH         ") != 0 &&
+                              strcmp(ID.c_str(), "PRECISION     ") != 0)) {
+    error(133);
   } else {
     INSYMBOL();
-    if (SY == LPARENT) {
+    if (SY == Symbol::LPARENT) {
       INSYMBOL();
     } else {
-      ERROR(9);
+      error(9);
     }
     su = bl->FSYS;
-    su[RPARENT] = true;
+    su[symbol::RPARENT] = true;
     EXPRESSION(bl, su, X);
-    if (X.TYP != INTS) {
-      ERROR(134);
+    if (X.types != Types::INTS) {
+      error(134);
     }
-    if (SY != RPARENT) {
-      ERROR(4);
+    if (SY != Symbol::RPARENT) {
+      error(4);
     }
-    if (strcmp(METHOD, "WIDTH         ") == 0) {
+    if (strcmp(METHOD.c_str(), "WIDTH         ") == 0) {
       EMIT1(30, 1);
     } else {
       EMIT1(30, 2);
@@ -876,40 +887,40 @@ void COUTMETHODS(BlockLocal *bl) {
 }
 
 void OUTPUTSTATEMENT(BlockLocal *bl) {
-  ITEM X;
-  SYMSET su;
+  Item X;
+  SymbolSet su;
   INSYMBOL();
-  if (SY == PERIOD) {
+  if (SY == Symbol::PERIOD) {
     COUTMETHODS(bl);
   } else {
     EMIT(89);
-    if (SY == OUTSTR) {
+    if (SY == Symbol::OUTSTR) {
       do {
         INSYMBOL();
-        if (SY == ENDLSY) {
+        if (SY == Symbol::ENDLSY) {
           EMIT(63);
           INSYMBOL();
-        } else if (SY == STRNG) {
+        } else if (SY == Symbol::STRNG) {
           EMIT1(24, SLENG);
           EMIT1(28, INUM);
           INSYMBOL();
         } else {
           su = bl->FSYS;
-          su[OUTSTR] = true;
-          su[SEMICOLON] = true;
+          su[symbol::OUTSTR] = true;
+          su[symbol::SEMICOLON] = true;
           EXPRESSION(bl, su, X);
           if (!(STANTYPS[X.TYP])) {
-            ERROR(41);
+            error(41);
           }
-          if (X.TYP == REALS) {
+          if (X.types == Types::REALS) {
             EMIT(37);
           } else {
-            EMIT1(29, (int)X.TYP);
+            EMIT1(29, static_cast<int>(X.types));
           }
         }
-      } while (SY == OUTSTR);
+      } while (SY == Symbol::OUTSTR);
     } else {
-      ERROR(128);
+      error(128);
     }
     EMIT(90);
   }
@@ -917,153 +928,181 @@ void OUTPUTSTATEMENT(BlockLocal *bl) {
 /* declared in SWITCHSTATEMENT therefore available to the contained functions
  * ala Pascal */
 struct SwitchLocal {
-  ITEM X;
+  Item X;
   int I, J, K, LC1, LC2, LC3;
   bool DEFOUND;
   struct {
-    INDEX VAL, LC;
-  } CASETAB[CSMAX + 1];
-  int EXITTAB[CSMAX + 1];
+    Index VAL, line_count;
+  } __attribute__((aligned(8))) CASETAB[CSMAX + 1];
+
+  std::array<int, CSMAX + 1> EXITTAB;
   BlockLocal *bl;
-};
-static void CASELABEL(SwitchLocal *);
-static void ONECASE(SwitchLocal *);
-void CASELABEL(SwitchLocal *sl) {
-  CONREC LAB = {NOTYP, 0l};
-  long K;
-  SYMSET su;
-  su = sl->bl->FSYS;
-  su[COLON] = true;
-  CONSTANT(sl->bl, su, LAB);
-  if (LAB.TP != sl->X.TYP) {
-    ERROR(47);
-  } else if (sl->I == CSMAX) {
-    FATAL(6);
+} __attribute__((aligned(128))) __attribute__((packed));
+
+static void CASELABEL(SwitchLocal * /*sl*/);
+static void ONECASE(SwitchLocal * /*sl*/);
+
+void CASELABEL(SwitchLocal *switch_local) {
+  CONREC LAB = {Types::NOTYP, 0L};
+  int64_t K = 0;
+  SymbolSet su;
+  su = switch_local->bl->FSYS;
+  su[symbol::COLON] = true;
+  CONSTANT(switch_local->bl, su, LAB);
+  if (LAB.TP != switch_local->X.types) {
+    error(47);
+  } else if (switch_local->I == CSMAX) {
+    fatal(6);
   } else {
-    sl->I = sl->I + 1;
+    switch_local->I = switch_local->I + 1;
     K = 0;
-    sl->CASETAB[sl->I].VAL = (int)LAB.I;
-    sl->CASETAB[sl->I].LC = LC;
+    switch_local->CASETAB[switch_local->I].VAL = static_cast<int>(LAB.I);
+    switch_local->CASETAB[switch_local->I].line_count = line_count;
     do {
       K = K + 1;
-    } while (sl->CASETAB[K].VAL != LAB.I);
-    if (K < sl->I) {
-      ERROR(1);
+    } while (switch_local->CASETAB[K].VAL != LAB.I);
+    if (K < switch_local->I) {
+      error(1);
     }
   }
 }
 
-void ONECASE(SwitchLocal *sl) {
-  SYMSET su;
+void ONECASE(SwitchLocal *switch_local) {
+  SymbolSet su;
   if (CONSTBEGSYS[SY]) {
-    CASELABEL(sl);
-    if (SY == COLON) {
+    CASELABEL(switch_local);
+    if (SY == Symbol::COLON) {
       INSYMBOL();
     } else {
-      ERROR(5);
+      error(5);
     }
-    su = sl->bl->FSYS;
-    su[CASESY] = true;
-    su[SEMICOLON] = true;
-    su[DEFAULTSY] = true;
-    su[RSETBRACK] = true;
+
+    su = switch_local->bl->FSYS;
+    su[symbol::CASESY] = true;
+    su[symbol::SEMICOLON] = true;
+    su[symbol::DEFAULTSY] = true;
+    su[symbol::RSETBRACK] = true;
+
     while (STATBEGSYS[SY] || ASSIGNBEGSYS[SY]) {
-      STATEMENT(sl->bl, su);
+      STATEMENT(switch_local->bl, su);
     }
+
     su = 0;
-    su[CASESY] = true;
-    su[DEFAULTSY] = true;
-    su[RSETBRACK] = true;
-    TEST(su, sl->bl->FSYS, 125);
+    su[symbol::CASESY] = true;
+    su[symbol::DEFAULTSY] = true;
+    su[symbol::RSETBRACK] = true;
+    TEST(su, switch_local->bl->FSYS, 125);
   }
 }
 
-void SWITCHSTATEMENT(BlockLocal *bl) {
+void SWITCHSTATEMENT(BlockLocal *block_local) {
   // int K;
-  struct SwitchLocal sl;
-  memset(&sl, 0, sizeof(struct SwitchLocal));
-  sl.bl = bl;
-  SYMSET su;
+  struct SwitchLocal switch_local {};
+  memset(&switch_local, 0, sizeof(struct SwitchLocal));
+
+  switch_local.bl = block_local;
+  SymbolSet su;
+
   BREAKPNT = BREAKPNT + 1;
   BREAKLOC[BREAKPNT] = 0;
   INSYMBOL();
-  if (SY == LPARENT) {
+
+  if (SY == Symbol::LPARENT) {
     INSYMBOL();
   } else {
-    ERROR(9);
+    error(9);
   }
-  sl.I = 0;
-  su = bl->FSYS;
-  su[RPARENT] = true;
-  su[COLON] = true;
-  su[LSETBRACK] = true;
-  su[CASESY] = true;
-  EXPRESSION(bl, su, sl.X);
-  if (!(sl.X.TYP == INTS || sl.X.TYP == BOOLS || sl.X.TYP == CHARS ||
-        sl.X.TYP == NOTYP)) {
-    ERROR(23);
+
+  switch_local.I = 0;
+  su = block_local->FSYS;
+  su[symbol::RPARENT] = true;
+  su[symbol::COLON] = true;
+  su[symbol::LSETBRACK] = true;
+  su[symbol::CASESY] = true;
+  EXPRESSION(block_local, su, switch_local.X);
+
+  if (switch_local.X.types != Types::INTS &&
+      switch_local.X.types != Types::BOOLS &&
+      switch_local.X.types != Types::CHARS &&
+      switch_local.X.types != Types::NOTYP) {
+    error(23);
   }
-  if (SY == RPARENT) {
+
+  if (SY == Symbol::RPARENT) {
     INSYMBOL();
   } else {
-    ERROR(4);
+    error(4);
   }
-  sl.LC1 = LC;
+
+  switch_local.LC1 = line_count;
   EMIT(12);
-  if (SY == LSETBRACK) {
+
+  if (SY == Symbol::LSETBRACK) {
     INSYMBOL();
   } else {
-    ERROR(106);
+    error(106);
   }
-  if (SY == CASESY) {
+  if (SY == Symbol::CASESY) {
     INSYMBOL();
   } else {
-    ERROR(124);
+    error(124);
   }
-  ONECASE(&sl);
-  while (SY == CASESY) {
+
+  ONECASE(&switch_local);
+
+  while (SY == Symbol::CASESY) {
     INSYMBOL();
-    ONECASE(&sl);
+    ONECASE(&switch_local);
   }
-  sl.DEFOUND = false;
-  if (SY == DEFAULTSY) {
+
+  switch_local.DEFOUND = false;
+  if (SY == Symbol::DEFAULTSY) {
     INSYMBOL();
-    if (SY == COLON) {
+
+    if (SY == Symbol::COLON) {
       INSYMBOL();
     } else {
-      ERROR(5);
+      error(5);
     }
-    sl.LC3 = LC;
+
+    switch_local.LC3 = line_count;
+
     while (STATBEGSYS[SY] || ASSIGNBEGSYS[SY]) {
-      su = bl->FSYS;
-      su[SEMICOLON] = true;
-      su[RSETBRACK] = true;
-      STATEMENT(bl, su);
+      su = block_local->FSYS;
+      su[symbol::SEMICOLON] = true;
+      su[symbol::RSETBRACK] = true;
+      STATEMENT(block_local, su);
     }
-    sl.DEFOUND = true;
+    switch_local.DEFOUND = true;
   }
-  sl.LC2 = LC;
+
+  switch_local.LC2 = line_count;
   EMIT(10);
-  CODE[sl.LC1].Y = LC;
-  for (sl.K = 1; sl.K <= sl.I; sl.K++) {
-    EMIT1(13, sl.CASETAB[sl.K].VAL);
-    EMIT1(13, sl.CASETAB[sl.K].LC);
+  code.at(switch_local.LC1).Y = line_count;
+
+  for (switch_local.K = 1; switch_local.K <= switch_local.I; switch_local.K++) {
+    EMIT1(13, switch_local.CASETAB[switch_local.K].VAL);
+    EMIT1(13, switch_local.CASETAB[switch_local.K].line_count);
   }
-  if (sl.DEFOUND) {
+
+  if (switch_local.DEFOUND) {
     EMIT2(13, -1, 0);
-    EMIT1(13, sl.LC3);
+    EMIT1(13, switch_local.LC3);
   }
+
   EMIT1(10, 0);
-  CODE[sl.LC2].Y = LC;
-  if (SYMCNT == 1) {
-    LOCATION[LNUM] = LC;
+  code.at(switch_local.LC2).Y = line_count;
+  if (symbol_count == 1) {
+    LOCATION[LNUM] = line_count;
   }
-  if (SY == RSETBRACK) {
+
+  if (SY == Symbol::RSETBRACK) {
     INSYMBOL();
   } else {
-    ERROR(104);
+    error(104);
   }
-  FIXBREAKS(LC);
+
+  FIXBREAKS(line_count);
 }
 
 } // namespace Cstar
